dynamic `manifest.json` based on `package.json` Fixes #53
