+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "sensasi-navbar-crx",
-  "description": "sensasi-navbar-crx",
+  "description": "Sensasi Navbar is a Chrome extension designed to enhance navigation within the browser, particularly when Chrome is in full-screen mode (F11). It provides a navigation bar that assists in navigating through tabs, offering convenience and ease of use.",
   "version": "0.1.0",
   "main": "index.js",
   "private": "true",
